Split project into different files.
